+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/parties/api/v3_1_8/GetAccountParty.kt
@@ -9,7 +9,6 @@ import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
 import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
 import com.forgerock.uk.openbanking.support.account.AccountFactory
 import com.forgerock.uk.openbanking.support.account.AccountRS
-import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_8
 import com.forgerock.uk.openbanking.tests.functional.account.access.BaseAccountApi3_1_8
 import uk.org.openbanking.datamodel.account.*
 
@@ -20,7 +19,7 @@ class GetAccountParty(version: OBVersion, tppResource: CreateTppCallback.TppReso
         val permissions = listOf(OBExternalPermissions1Code.READPARTY, OBExternalPermissions1Code.READACCOUNTSDETAIL)
         val (_, accessToken) = accountAccessConsentApi.createConsentAndGetAccessToken(permissions)
         val (accountId, psuId) = AccountRS().getFirstAccountIdAndPsuId(
-            accountAndTransaction3_1_8.Links.links.GetAccounts,
+            accountsApiLinks.GetAccounts,
             accessToken
         )
         assertThat(accountId).isNotEqualTo("")
@@ -30,7 +29,7 @@ class GetAccountParty(version: OBVersion, tppResource: CreateTppCallback.TppReso
         // When
         val result = AccountRS().getAccountsDataEndUser<OBReadParty2>(
             AccountFactory.urlWithAccountId(
-                accountAndTransaction3_1_8.Links.links.GetAccountParty,
+                accountsApiLinks.GetAccountParty,
                 accountId
             ), accessToken
         )
