+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/products/api/v3_1_8/GetProducts.kt
@@ -1,0 +1,29 @@
+package com.forgerock.uk.openbanking.tests.functional.account.products.api.v3_1_8
+
+import assertk.assertThat
+import assertk.assertions.isNotEmpty
+import assertk.assertions.isNotNull
+import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
+import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
+import com.forgerock.uk.openbanking.support.account.AccountRS
+import com.forgerock.uk.openbanking.tests.functional.account.access.BaseAccountApi3_1_8
+import uk.org.openbanking.datamodel.account.*
+
+class GetProducts(version: OBVersion, tppResource: CreateTppCallback.TppResource): BaseAccountApi3_1_8(version, tppResource) {
+    fun shouldGetProductsTest() {
+        // Given
+        val permissions = listOf(
+            OBExternalPermissions1Code.READACCOUNTSDETAIL, OBExternalPermissions1Code.READPRODUCTS
+        )
+        val (_, accessToken) = accountAccessConsentApi.createConsentAndGetAccessToken(permissions)
+
+        // When
+        val result = AccountRS().getAccountsData<OBReadProduct2>(
+            accountsApiLinks.GetProducts, accessToken
+        )
+
+        // Then
+        assertThat(result).isNotNull()
+        assertThat(result.data.product).isNotEmpty()
+    }
+}
