+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/products/legacy/LegacyGetAccountProductTest.kt
@@ -1,4 +1,4 @@
-package com.forgerock.uk.openbanking.tests.functional.account.products
+package com.forgerock.uk.openbanking.tests.functional.account.products.legacy
 
 import assertk.assertThat
 import assertk.assertions.isNotEmpty
@@ -10,13 +10,12 @@ import com.forgerock.uk.openbanking.support.account.AccountAS
 import com.forgerock.uk.openbanking.support.account.AccountFactory.Companion.urlWithAccountId
 import com.forgerock.uk.openbanking.support.account.AccountRS
 import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_2
-import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_8
 import org.junit.jupiter.api.Test
 import uk.org.openbanking.datamodel.account.*
 import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code.READACCOUNTSDETAIL
 import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code.READPRODUCTS
 
-class GetAccountProductTest(val tppResource: CreateTppCallback.TppResource) {
+class LegacyGetAccountProductTest(val tppResource: CreateTppCallback.TppResource) {
 
     @EnabledIfVersion(
         type = "accounts",
@@ -54,52 +53,6 @@ class GetAccountProductTest(val tppResource: CreateTppCallback.TppResource) {
         val result = AccountRS().getAccountsData<OBReadProduct2>(
             urlWithAccountId(
                 accountAndTransaction3_1_2.Links.links.GetAccountProduct,
-                accountId
-            ), accessToken
-        )
-
-        // Then
-        assertThat(result).isNotNull()
-        assertThat(result.data.product).isNotEmpty()
-    }
-
-
-    @EnabledIfVersion(
-        type = "accounts",
-        apiVersion = "v3.1.8",
-        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetAccountProduct"],
-        apis = ["products"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5", "v.3.1.4", "v.3.1.3"]
-    )
-    @Test
-    fun shouldGetAccountProduct_v3_1_8() {
-        // Given
-        val consentRequest = OBReadConsent1().data(
-            OBReadData1()
-                .permissions(
-                    listOf(
-                        READACCOUNTSDETAIL, READPRODUCTS
-                    )
-                )
-        )
-            .risk(OBRisk2())
-        val consent = AccountRS().consent<OBReadConsentResponse1>(
-            accountAndTransaction3_1_8.Links.links.CreateAccountAccessConsent,
-            consentRequest,
-            tppResource.tpp
-        )
-        val accessToken = AccountAS().getAccessToken(
-            consent.data.consentId,
-            tppResource.tpp.registrationResponse,
-            psu,
-            tppResource.tpp
-        )
-        val accountId = AccountRS().getFirstAccountId(accountAndTransaction3_1_8.Links.links.GetAccounts, accessToken)
-
-        // When
-        val result = AccountRS().getAccountsData<OBReadProduct2>(
-            urlWithAccountId(
-                accountAndTransaction3_1_8.Links.links.GetAccountProduct,
                 accountId
             ), accessToken
         )
